requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2.0
psycopg2-binary>=2.9
python-decouple>=3.8
